refactor(harmony): bring up @expo-harmony/metro-config
