+++ package.json
@@ -14,7 +14,7 @@
     "prebuild:android": "expo prebuild --platform android --clean",
     "prebuild:ios": "expo prebuild --platform ios --clean",
     "prebuild:harmony": "node scripts/harmony/prebuild.js",
-    "oh": "react-native start --config metro.harmony.config.js --port 8082",
+    "oh": "EXPO_METRO_TARGET=harmony expo start --port 8082 --localhost",
     "oh:install": "node scripts/harmony/prepare-dependencies.js && cd harmony && ohpm install --all",
     "oh:bundle": "node scripts/harmony/bundle.js",
     "oh:build": "yarn prebuild:harmony && yarn oh:install && yarn oh:bundle && cd harmony && hvigorw assembleHap --mode module -p product=default -p buildMode=debug --no-daemon",
@@ -90,6 +90,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.29.6",
+    "@expo-harmony/metro-config": "^0.0.1",
     "@react-native-community/cli": "20.0.0",
     "@react-native/eslint-config": "^0.79.7",
     "@react-native/metro-config": "^0.86.2",
@@ -102,6 +103,7 @@
     "eslint-plugin-prettier": "^5.5.1",
     "jest": "^29.2.1",
     "jest-expo": "~56.0.5",
+    "metro-config": "^0.84.3",
     "openapi-ts-request": "^0.13.3",
     "prettier": "^3.3.3",
     "prettier-plugin-organize-imports": "^4.1.0",
